requirements.txt
PyYAML >= 3.11
ansible >= 2.16.14
#Modules for python script
GitPython
ansible-runner